refactor: reformatted script into python package layout
