+++ src/evilfonttool/__init__.py
@@ -1,0 +1,8 @@
+from evilfonttool._core import (
+    createfonts,
+    createstealthfont,
+    createhtml,
+    create_doc,
+)
+
+__all__ = ["createfonts", "createstealthfont", "createhtml", "create_doc"]
