+++ src/evilfonttool/__main__.py
@@ -1,0 +1,3 @@
+from evilfonttool.cli import main
+
+main()
