+++ src/evilfonttool/data/template.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+
+<link rel="stylesheet" href="fonts.css">
+
+
+<html>
+<body>
+
+<!-- #STUFF HERE -->
+
+</body>
+</html>
